requirements.txt
matplotlib
gdax
numpy
requests
pyyaml
